search: drop EnqueueRequestForSearchOwnerMultiCluster, use map-fn
